requirements.txt
aiofiles==24.1.0
aiohttp==3.12.6
aiohttp-cors==0.7.0
aiomysql==0.2.0
asyncmy==0.2.10
beautifulsoup4==4.13.4
discord.py==2.6.3
dnspython==2.6.1
emoji==2.2.0
greenlet==3.2.2
logtail-python==0.3.3
lxml==5.4.0
lyricsgenius==3.6.2
pillow==11.2.1
psutil==7.0.0
PyNaCl==1.5.0
pyrebase4==4.8.0
pytz
requests~=2.32.3
sqlalchemy==2.0.41